adds crontroller test
